secure-script-runner/api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
celery>=5.3.4
redis>=5.0.1
boto3>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
requests>=2.31.0
docker>=6.1.0
httpx>=0.27.0
